requirements.txt
torch
numpy
matplotlib
tqdm
Pillow
rasterio
torchvision